refactor. move model out of api
