+++ src/constants.ts
@@ -4,3 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 export const EXTENSION_ID = 'GitHub.vscode-pull-request-github';
+export const VSLS_REQUEST_NAME = 'git';
+export const VSLS_GIT_PR_SESSION_NAME = 'ghpr';
+export const VSLS_REPOSITORY_INITIALIZATION_NAME = 'initialize';
+export const VSLS_STATE_CHANGE_NOFITY_NAME = 'statechange';
