--- src/git/local.ts
@@ -1,9 +1,0 @@
-
-import * as vscode from 'vscode';
-import { GitExtension } from '../typings/git';
-
-export function getAPI() {
-	const gitExtension = vscode.extensions.getExtension<GitExtension>('vscode.git').exports;
-	const git = gitExtension.getAPI(1);
-	return git;
-}
